DonutChartViewKeyItem moved to sample-compose-fragment
